+++ aldrin/src/lib.rs
@@ -89,7 +89,9 @@ pub use channel::{
     UnclaimedReceiver, UnclaimedSender,
 };
 pub use client::Client;
-pub use discoverer::{Discoverer, DiscovererBuilder, DiscovererEvent, DiscovererEventKind};
+pub use discoverer::{
+    Discoverer, DiscovererBuilder, DiscovererEntry, DiscovererEvent, DiscovererEventKind,
+};
 pub use error::Error;
 pub use handle::Handle;
 pub use lifetime::{Lifetime, LifetimeId, LifetimeScope};