+++ aldrin/CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `Discoverer::object_id()` and `service_id()` can be used to query known objects and
   services. Previously, this was only possible through the short-lived events.
+- `Discoverer::entry()` can be used to query objects and services that correspond to one specific
+  key.
 
 ### Changed